+++ R/htest_tidiers.R
@@ -1,16 +1,18 @@
 #' Tidying methods for an htest object
 #'
-#' Tidies hypothesis test objects, such as those from \code{cor.test},
-#' \code{t.test}, and \code{wilcox.test}, into a one-row data frame.
+#' Tidies hypothesis test objects, such as those from 
+#' [stats::cor.test()], [stats::t.test()],
+#' [stats::wilcox.test()], and [stats::chisq.test()],
+#' into a one-row data frame.
 #'
-#' @details No \code{augment} method is provided for \code{"htest"},
-#' since there is no sense in which a hypothesis test generates one
-#' value for each observation.
+#' @details `augment` method is defined only for chi-squared tests,
+#' since there is no sense, for other tests, in which a hypothesis test
+#' generates one value for each observation.
 #'
-#' @param x An object of class \code{"htest"}
+#' @param x An object of class `"htest"`
 #' @param ... extra arguments (not used)
 #'
-#' @return Both \code{tidy} and \code{glance} return the same output,
+#' @return Both `tidy` and `glance` return the same output,
 #' a one-row data frame with one or more of the following columns:
 #'   \item{estimate}{Estimate of the effect size}
 #'   \item{statistic}{Test statistic used to compute the p-value}
@@ -28,6 +30,19 @@
 #'
 #' Which columns are included depends on the hypothesis test used.
 #'
+#' For chi-squared tests, `augment.htest`  will returns, for each
+#' cell of the tested table, the additional columns:
+#'   \item{.observed}{Observed count}
+#'   \item{.prop}{Proportion of the total}
+#'   \item{.row.prop}{Row proportion (2 dimensions table only)}
+#'   \item{.col.prop}{Column proportion (2 dimensions table only)}
+#'   \item{.expected}{Expected count under the null hypothesis}
+#'   \item{.residuals}{Pearson residual}
+#'   \item{.stdres}{Standardized residual}
+#'
+#' See [stats::chisq.test()] for more details on
+#' how residuals are computed.
+#'
 #' @examples
 #'
 #' tt <- t.test(rnorm(10))
@@ -37,11 +52,15 @@
 #' tt <- t.test(mpg ~ am, data = mtcars)
 #' tidy(tt)
 #'
-#' wt <- wilcox.test(mpg ~ am, data = mtcars)
+#' wt <- wilcox.test(mpg ~ am, data = mtcars, conf.int = TRUE, exact = FALSE)
 #' tidy(wt)
 #'
 #' ct <- cor.test(mtcars$wt, mtcars$mpg)
 #' tidy(ct)
+#'
+#' chit <- chisq.test(xtabs(Freq ~ Sex + Class, data = as.data.frame(Titanic)))
+#' tidy(chit)
+#' augment(chit)
 #'
 #' @name htest_tidiers
 NULL
@@ -95,3 +114,42 @@ tidy.htest <- function(x, ...) {
 #' @rdname htest_tidiers
 #' @export
 glance.htest <- function(x, ...) tidy(x)
+
+#' @rdname htest_tidiers
+#' @export
+augment.htest <- function(x, ...) {
+  if (all(c("observed", "expected", "residuals", "stdres") %in% names(x))) {
+    return(augment_chisq_test(x, ...))
+  } else {
+    stop(
+      "augment doesn't know how to deal with class htest with method '",
+      x$method, "'"
+    )
+  }
+}
+
+augment_chisq_test <- function(x, ...) {
+  d <- length(dimnames(as.table(x$observed)))
+  ret <- as.data.frame(as.table(x$observed))
+  names(ret)[d + 1] <- ".observed"
+
+  ret <- cbind(
+    ret,
+    .prop = as.data.frame(prop.table(as.table(x$observed)))[[d + 1]]
+  )
+  if (d == 2) {
+    ret <- cbind(
+      ret,
+      .row.prop = as.data.frame(prop.table(as.table(x$observed), 1))[[d + 1]]
+    )
+    ret <- cbind(
+      ret,
+      .col.prop = as.data.frame(prop.table(as.table(x$observed), 2))[[d + 1]]
+    )
+  }
+
+  ret <- cbind(ret, .expected = as.data.frame(as.table(x$expected))[[d + 1]])
+  ret <- cbind(ret, .residuals = as.data.frame(as.table(x$residuals))[[d + 1]])
+  ret <- cbind(ret, .stdres = as.data.frame(as.table(x$stdres))[[d + 1]])
+  ret
+}