+++ R/smooth.spline_tidiers.R
@@ -3,7 +3,7 @@
 #' This combines the original data given to smooth.spline with the
 #' fit and residuals
 #'
-#' @details No \code{tidy} method is provided for smooth.spline objects.
+#' @details No `tidy` method is provided for smooth.spline objects.
 #'
 #' @param x a smooth.spline object
 #' @param data defaults to data used to fit model
@@ -24,7 +24,7 @@
 
 #' @rdname smooth.spline_tidiers
 #'
-#' @return \code{augment} returns the original data with extra columns:
+#' @return `augment` returns the original data with extra columns:
 #'   \item{.fitted}{Fitted values of model}
 #'   \item{.resid}{Residuals}
 #'
@@ -42,9 +42,9 @@ augment.smooth.spline <- function(x, data = x$data, ...) {
 
 #' @rdname smooth.spline_tidiers
 #'
-#' @return \code{glance} returns one row with columns
+#' @return `glance` returns one row with columns
 #'   \item{spar}{smoothing parameter}
-#'   \item{lambda}{choice of lambda corresponding to \code{spar}}
+#'   \item{lambda}{choice of lambda corresponding to `spar`}
 #'   \item{df}{equivalent degrees of freedom}
 #'   \item{crit}{minimized criterion}
 #'   \item{pen.crit}{penalized criterion}