+++ tests/testthat/test-fitdistr.R
@@ -1,6 +1,7 @@
 context("fitdistr tidiers")
 
 test_that("fitdistr tidiers work", {
+  skip_if_not_installed("MASS")
   set.seed(2015)
   x <- rnorm(100, 5, 2)
   fit <- suppressWarnings(MASS::fitdistr(x, dnorm, list(mean = 3, sd = 1)))